--- __tests__/html/fluentTheme/preChatMessageActivity.js
@@ -1,5 +1,0 @@
-/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
-
-describe('Fluent theme applied', () => {
-  test('should display pre-chat message', () => runHTML('fluentTheme/preChatMessageActivity'));
-});
