+++ __tests__/html2/fluentTheme/preChatMessageActivity/layout.wide.html
@@ -15,6 +15,8 @@
     <main id="webchat"></main>
     <script type="text/babel">
       run(async function () {
+        await host.windowSize(720, 640, document.getElementById('webchat'));
+
         const {
           React,
           ReactDOM: { render },
@@ -104,23 +106,7 @@
 
         await pageConditions.numActivitiesShown(1);
 
-        await host.snapshot();
-
-        // WHEN: The first card action button is clicked.
-        await host.click(
-          document.querySelector(`[data-testid="${WebChat.testIds.preChatMessageActivityStarterPromptsCardAction}"]`)
-        );
-
-        // THEN: Send box text box should have the value typed for the user.
-        await host.snapshot();
-
-        // WHEN: The sixth (last) card action button is clicked.
-        await host.click(
-          document.querySelectorAll(`[data-testid="${WebChat.testIds.preChatMessageActivityStarterPromptsCardAction}"]`)[5]
-        );
-
-        // THEN: Send box text box should be replaced.
-        await host.snapshot();
+        await host.snapshot('local');
       });
     </script>
   </body>
