+++ __tests__/html2/fluentTheme/preChatMessageActivity/layout.wide.html.snap-1.png
