+++ __tests__/html2/fluentTheme/preChatMessageActivity/layout.withoutImage.html
@@ -92,7 +92,7 @@
 
         await pageConditions.numActivitiesShown(1);
 
-        await host.snapshot();
+        await host.snapshot('local');
       });
     </script>
   </body>
