+++ __tests__/html2/fluentTheme/preChatMessageActivity/layout.withoutImage.html.snap-1.png
