+++ __tests__/html2/fluentTheme/preChatMessageActivity/simple.html.snap-1.png
