+++ docker-compose-wsl2.yml
@@ -42,6 +42,7 @@ services:
     stop_grace_period: 0s
     volumes:
       - ./__tests__/html/:/var/web/__tests__/html/
+      - ./__tests__/html2/:/var/web/__tests__/html2/
       - ./packages/bundle/dist/:/var/web/packages/bundle/dist/
       - ./packages/fluent-theme/dist/:/var/web/packages/fluent-theme/dist/
       - ./packages/test/fluent-bundle/dist/:/var/web/packages/test/fluent-bundle/dist/
